+++ CHANGES.md
@@ -5,6 +5,9 @@ that violated one-to-one correspondance between a red side emission feature and
 its 'duplicate' on the blue side.
 - The BackgroundSubtractSpectrum stage no longer adds the background counts
 in quadrature to the error in the 2d spectrum.
+- RuntimeContext raises a custom attribute error if an attribute is missing, this
+error suggests to the user that the attribute name is likely missing from the configuration
+file.
 
 0.1.2 (2019-12-06)
 -------------------